k8s: change client.Get to use an ObjectReference
