+++ internal/engine/event_watch_manager.go
@@ -2,7 +2,6 @@ package engine
 
 import (
 	"context"
-	"regexp"
 	"sync"
 	"time"
 
@@ -73,31 +72,6 @@ func (m *EventWatchManager) OnChange(ctx context.Context, st store.RStore) {
 	go m.dispatchEventsLoop(ctx, ch, st)
 }
 
-// Tests whether a string is a valid version for a k8s resource type.
-// from https://kubernetes.io/docs/tasks/access-kubernetes-api/custom-resources/custom-resource-definition-versioning/#version-priority
-// Versions start with a v followed by a number, an optional beta or alpha designation, and optional additional numeric
-// versioning information. Broadly, a version string might look like v2 or v2beta1.
-var versionRegex = regexp.MustCompile(`^v\d+(?:(?:alpha|beta)(?:\d+)?)?$`)
-
-func getGroup(involvedObject v1.ObjectReference) string {
-	// For some types, APIVersion is incorrectly just the group w/ no version, which leads GroupVersionKind to return
-	// a value where Group is empty and Version contains the group, so we need to correct for that.
-	// An empty Group is valid, though: it's empty for apps in the core group.
-	// So, we detect this situation by checking if the version field is valid.
-
-	// this stems from group/version not necessarily being populated at other points in the API. see more info here:
-	// https://github.com/kubernetes/client-go/issues/308
-	// https://github.com/kubernetes/kubernetes/issues/3030
-
-	gvk := involvedObject.GroupVersionKind()
-	group := gvk.Group
-	if !versionRegex.MatchString(gvk.Version) {
-		group = involvedObject.APIVersion
-	}
-
-	return group
-}
-
 func (m *EventWatchManager) createEntry(ctx context.Context, involvedObject v1.ObjectReference) uidMapEntry {
 	ret := uidMapEntry{
 		resourceVersion:   involvedObject.ResourceVersion,
@@ -105,14 +79,7 @@ func (m *EventWatchManager) createEntry(ctx context.Context, involvedObject v1.O
 		expiresAt:         m.clock.Now().Add(uidMapEntryTTL),
 	}
 
-	o, err := m.kClient.Get(
-		getGroup(involvedObject),
-		"", // version is taken care of by DiscoveryRESTMapper
-		involvedObject.Kind,
-		involvedObject.Namespace,
-		involvedObject.Name,
-		involvedObject.ResourceVersion,
-	)
+	o, err := m.kClient.GetByReference(involvedObject)
 	if err != nil {
 		// if the lookup was an error, wipe out resourceVersion so that we don't cache a potentially
 		// ephemeral negative result
